Move general bit operation utilities to BitUtilities in Eutherion.Shared. Add new SetBits() method.
