+++ Sandra.Chess/ChessTypes.cs
@@ -2,7 +2,7 @@
 /*********************************************************************************
  * ChessTypes.cs
  *
- * Copyright (c) 2004-2020 Henk Nicolai
+ * Copyright (c) 2004-2021 Henk Nicolai
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -21,7 +21,7 @@
 
 using Sandra.Chess;
 using System.Collections.Generic;
-using System.Runtime.CompilerServices;
+using System.Numerics;
 
 namespace Sandra.Chess
 {
@@ -138,50 +138,9 @@ namespace Sandra
         public static ulong ToVector(this Square square) => 1UL << (int)square;
 
         /// <summary>
-        /// Gets the index of the single bit of the vector, or an undefined value if the number of set bits in the vector is not equal to one.
-        /// </summary>
-        public static int GetSingleBitIndex(this ulong oneBitVector)
-        {
-            // Constant masks.
-            const ulong m1 = 0x5555555555555555;  // 010101010101...
-            const ulong m2 = 0x3333333333333333;  // 001100110011...
-            const ulong m4 = 0x0f0f0f0f0f0f0f0f;  // 000011110000...
-            const ulong m8 = 0x00ff00ff00ff00ff;
-            const ulong m16 = 0x0000ffff0000ffff;
-            const ulong m32 = 0x00000000ffffffff;
-
-            // Calculate the index of the single set bit by testing it against several predefined constants.
-            // This index is built as a binary value.
-            int index = ((oneBitVector & m32) == 0 ? 32 : 0) |
-                        ((oneBitVector & m16) == 0 ? 16 : 0) |
-                        ((oneBitVector & m8) == 0 ? 8 : 0) |
-                        ((oneBitVector & m4) == 0 ? 4 : 0) |
-                        ((oneBitVector & m2) == 0 ? 2 : 0) |
-                        ((oneBitVector & m1) == 0 ? 1 : 0);
-
-            return index;
-        }
-
-        /// <summary>
-        /// Enumerates all indices for which the corresponding bit in the vector is true.
-        /// </summary>
-        public static IEnumerable<int> Indices(this ulong vector)
-        {
-            while (vector != 0)
-            {
-                // Select the least significant 1-bit using a well known trick.
-                ulong oneBit = vector & (0U - vector);
-                yield return GetSingleBitIndex(oneBit);
-
-                // Zero the least significant 1-bit so the index of the next 1-bit can be yielded.
-                vector ^= oneBit;
-            }
-        }
-
-        /// <summary>
         /// Gets the single square for which this vector is set, or an undefined value if the number of squares in the vector is not equal to one.
         /// </summary>
-        public static Square GetSingleSquare(this ulong oneBitVector) => (Square)GetSingleBitIndex(oneBitVector);
+        public static Square GetSingleSquare(this ulong oneBitVector) => (Square)BitOperations.Log2(oneBitVector);
 
         /// <summary>
         /// Enumerates all squares for which this vector is set.
@@ -193,24 +152,6 @@ namespace Sandra
                 yield return ((Square)index);
             }
         }
-
-        /// <summary>
-        /// Tests if a vector has any set bits.
-        /// </summary>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static bool Test(this ulong vector) => vector != 0;
-
-        /// <summary>
-        /// Tests if a vector is equal to zero or a power of two, i.e. is true for zero or one bits exactly.
-        /// </summary>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static bool IsMaxOneBit(this ulong vector) => !vector.Test(vector - 1);
-
-        /// <summary>
-        /// Tests if another vector has any bits in common with this one.
-        /// </summary>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static bool Test(this ulong vector, ulong otherVector) => (vector & otherVector) != 0;
 
         /// <summary>
         /// Returns if the given <see cref="MoveCheckResult"/> represents a legal move, even if that move is incomplete.
