+++ Sandra.Chess/Game.cs
@@ -22,6 +22,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 
 namespace Sandra.Chess
 {
