+++ Sandra.Chess/Pgn/PgnSymbolStateMachine.cs
@@ -2,7 +2,7 @@
 /*********************************************************************************
  * PgnSymbolStateMachine.cs
  *
- * Copyright (c) 2004-2020 Henk Nicolai
+ * Copyright (c) 2004-2021 Henk Nicolai
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -20,6 +20,7 @@
 #endregion
 
 using System.Linq;
+using System.Numerics;
 
 namespace Sandra.Chess.Pgn
 {
