+++ Sandra.UI/PgnEditorExtensions.cs
@@ -2,7 +2,7 @@
 /*********************************************************************************
  * PgnEditorExtensions.cs
  *
- * Copyright (c) 2004-2020 Henk Nicolai
+ * Copyright (c) 2004-2021 Henk Nicolai
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -24,6 +24,7 @@ using Sandra.Chess;
 using Sandra.Chess.Pgn;
 using System;
 using System.Linq;
+using System.Numerics;
 
 namespace Sandra.UI
 {
